refactored mentee/mentor routes to user route
