--- pages/api/courses/mentors/[mentor_id].js
@@ -1,27 +1,0 @@
-// GET: /pages/api/courses/mentors/[mentor_id].js
-
-import { db } from '../../../../utils/api/firebase.config';
-import {
-  collection,
-  doc,
-  getDoc,
-  getDocs,
-  addDoc,
-  query,
-  where
-} from "firebase/firestore";
-
-export default async function getCoursesByMentorId(req, res) {
-  const { mentor_id } = req.query;
-  try {
-    const q = query(collection(db, 'courses'), where('mentor.id', '==', Number(mentor_id)));
-    const result = [];
-    const querySnapshot = await getDocs(q);
-    querySnapshot.forEach(doc => {
-      result.push(doc.data());
-    });
-    res.status(200).json(result);
-  } catch (err) {
-    res.status(400).send(`Error retrieving courses by mentor id: ${err}`);
-  }
-}
