+++ pages/api/users/[user_id].js
@@ -1,11 +1,13 @@
-// GET: /pages/api/mentors/[mentor_id].js
+// GET: /pages/api/users/[user_id].js
+
 import { db } from '../../../utils/api/firebase.config';
 import { collection, doc, getDoc } from "firebase/firestore";
 
-export default async function getMentorInfo(req, res) {
-  const { mentor_id } = req.query;
+export default async function getUserInfo(req, res) {
+  const { user_id } = req.query;
+  console.log('uid: ', user_id);
   try {
-    const docRef = doc(db, 'mentors', mentor_id);
+    const docRef = doc(db, 'users', user_id);
     const docSnap = await getDoc(docRef);
     res.status(200).json(docSnap.data());
   } catch (err) {
